refactored templates to private package as identified in PR feedback
